commented out tests for tcp_client
I added some tests for tcp_client but was unable to get things to compile.
Seems like maybe an issue with TcpClient.connect(..) and perhaps the way we use closure.
Everything compiles fine if I remove the call to connect().
Leaving this here commented out so that we can come back to it.